requirements.txt
# ==============================
# Core Web Framework
# ==============================
Flask==2.3.3
Werkzeug==3.1.4

# ==============================
# Environment & HTTP
# ==============================
python-dotenv==1.0.1
requests==2.31.0

# ==============================
# PDF Processing
# ==============================
pypdf==4.0.1

# ==============================
# OCR (for scanned PDFs)
# ==============================
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# ==============================
# LangChain (Pinned to avoid LangGraph conflicts)
# ==============================
langchain==0.1.17
langchain-core==0.1.53
langchain-community==0.0.38
langsmith==0.1.147

# ==============================
# Google Gemini
# ==============================
langchain-google-genai==0.0.11
google-genai==1.55.0
google-generativeai==0.4.1
